api-ref: document sorting parameters for networks
API users can use the 'sort_key' and 'sort_dir' to do the sorting
on list result. This patch improves the documentation of these two
sorting parameters.

In particular, we explicitly list all the attributes that can be used
as a sort_key for network resource. This is necessary because the list
of sort_key varies across different resources and it is hard for
API users to guess which resource support which sort_key.

Change-Id: I5183bfe4c835dae0f0bfae61114e9b9be2877b0f