fix vpnaas subnet_list_or_none type
The validate_subnet_list_or_none validator was rehomed into lib as part
of [1] and registered with the type of 'subnet_list_or_none'. While this
is a good type name, vpnaas already adds a validator with that type in
their master branch today [2]. As a result, running with neutron-lib
master can fail due to re-registration [3]. The way we've addressed this
in the past is to use a new type name in neutron-lib and then just
ensure the new type is used upon consumption.

This patch renames the validation type subnet_list_or_none to
list_of_subnets_or_none to address the issue [3].

[1] https://review.openstack.org/#/c/439972/
[2] https://github.com/openstack/neutron-vpnaas/blob/master/
neutron_vpnaas/extensions/vpnaas.py#L177
[3] http://logs.openstack.org/periodic/git.openstack.org/openstack/
vmware-nsx/master/openstack-tox-py35-with-neutron-lib-master/227d2f0/
job-output.txt.gz#_2018-04-23_06_17_55_841555

Change-Id: If1713edb26fd923047f2f6c022f0557b2f3ec41e