Add attribute propagate_uplink_status to port
This attribute indicates if the state of the port will follow
the state of its parent port. The first use case of this attribute
is SR-IOV. If it is a VF port, setting this attribute can make
the VF link state follow the state of the PF. In the future,
this attribute might be reused for trunk/subports as well.

Change-Id: If622b8df6b890530adace4821a6408ef942cc80a
Partial-Bug: #1722720