tell flake8 to ignore unicode
There is no builtin type 'unicode' under python 3. The code is already
doing the correct thing at runtime, but that check is not evaluated
when the linter runs under python 3 so it fails because the name is
not defined. This change tells the linter to ignore the line in the
exceptions package.

Change-Id: Idea04659d3699ab2243f6c325d13d43abe9b447a
Signed-off-by: Doug Hellmann <doug@doughellmann.com>