Placement client: always return body
In a few POST and PUT calls we didn't return the JSON body of the
response we got. By that we were just throwing away information we
already waited for.

This patch makes sure that if the HTTP response had a body we return it.

We leave one exception: PUT /resource_classes/ microversions 1.2 - 1.6.
Usage of that is strongly discouraged and starting from 1.7 the response
no longer has a body.

Change-Id: Ifa4db6edd4c656023039f7a68c09bf2cea508231