Use authorize instead of enforce in policy
the policy module was doing a manual check to see if the specific rule
was actually part of the rules that the enforcer is using. oslo.policy
already has a function that does just this, which is 'authorize'. That
will check the registered rules and raise an exception if that's not
fulfilled.

Change-Id: I9f04f8b8770b15ac24f9f1cd57a58c7e98b24d48