Fix random_mac_generator to make proper EUI64s
The initial implementation of random_mac_generator created MAC addresses
that were one octet too long - the MAC addresses would have seven octets
instead of the proper EUI64-standard six.

This resolved testing issues in the consuming patch in Neutron (see
the change referenced in the Needed-By line).  A test is added that
replicates the conditions of the failure.  In the failed tests, the base
mac '12:34:56:78' was instead parsed so it became '12:34:56:7:8'; when
the code went to add 2 octets to the result, there were too many octets
to form a valid EUI64 MAC address.

Implements: blueprint speed-up-neutron-bulk-creation
Needed-By: https://review.openstack.org/584061
Change-Id: I9a2c40709db9204c7686500deac9ee8e8471b1bc