Placement client: do not swallow exceptions
Placement client unconditionally swallowed many exceptions. I cannot
tell why that was ever considered a good idea. I can imagine some
users wanting to ignore some of those error conditions, but it should
be the responsibility of the user to decide. The client should not
unconditionally ignore all errors, but leave this choice for the user
of the client.

Particularly this placement client will be used by later patches for
minimum-bandwidth-allocation-placement-api where we do want to know
about some of these errors and retry placement operations when they
fail.

Change-Id: Ia14ba8cb273166fe3075328229f30d0cebae0480
Related-Bug: #1578989