Fix default value of subnet attribute for subnet onboard
In its current state, the subnet onboard extension forces callers
to pass a value for the subnet attribute extension. This is not
required when performing operations on a subnet pool, but the
validator forces a value to be sent on the request. This change
simply adjusts the default value of the subnet attribute in the
subnet onboard extension so that an empty list is assumed. This
allows for backward compatible operation of the subnetpool API,
allowing callers to choose not to pass this attribute when it
is not needed in a request.

Change-Id: I6c0b5aa784a380502b9e5fd062dacd10a95b4cbf
Partially-Implements: blueprint subnet-onboard