Add smartnic port type to VNIC_TYPES
Neutron OVS ML2 mechanism driver will allow binding the `smart-nic`
vnic_type. The rationale for creating new vnic_type and not using the
barmetal one is that there is a wide range of mechanism drivers that
use heirachical port binding for configuring TOR switches and we want
to allow this to work with smart NICs.

Story: #2003346
Change-Id: I91f63810626ce4e054e358f5de5e46434c4da131