New agent attribute: resources_synced
Agents supporting the guaranteed minimum bandwidth feature need to share
their resource view with neutron-server and in turn with Placement too.
The resource information is synchronized to neutron-server via the
periodic agent heartbeat therefore transient synchronization errors are
fixed by the next heartbeat. But synchronization to Placement is not
done periodically, but on a (mostly) on demand basis. Therefore to fix
transient errors of the synchronization to Placement we must remember
the success/failure of the last synchronization attempt.

This change introduces a new boolean agent attribute: resources_synced

This attribute is:
* admin-only (just like the whole agent extension)
* read-only (it is only updated by the internal synchronization process)

If this attribute is not set then no synchronization to Placement ever
happened (which is the natural state of agents not tracking their
resources via Placement).

If it is False, then another successful synchronization is needed before
the contents of Placement can be considered up to date.

If it is True, then this agent's resources are up to date in Placement.

Change-Id: Ia3dea82fabe5e068192fbe931db8d341d3d37149
Needed-By: https://review.openstack.org/630999
Partial-Bug: #1578989
See-Also: https://review.openstack.org/502306 (nova spec)
See-Also: https://review.openstack.org/508149 (neutron spec)