Delete floating IPs on network turned internal
The neutron-lib part of this change: https://review.openstack.org/624751

This change introduces a new shim extension
(floatingip-autodelete-internal) to signal a change in API behavior
introduced by the fix of a bug referred below. That is now we autodelete
unused floating IPs when changing a network to router:external=False.

It also documents the API change.

Change-Id: I247106e95abb5d6822aa35ae793531a1d79edbd4
Needed-By: https://review.openstack.org/624751
Partial-Bug: #1806032