Increase tempest-full jobs timeout to 3h
Jobs tempest-full and tempest-full-py3 are included in
neutron-lib's check and gate queues because they are in
lib-forward-testing and lib-forward-testing-python3 templates.
But in those templates those jobs have 2h timeout which is sometimes
not enough to complete job.

In Neutron repo we increased this timeout to 3h already some time ago
and that helped a lot with such timeouts so lets increase it also
in neuton-lib repo to avoid some rechecks.

Change-Id: Ic3ce87a8c3417957afe7fc1e1ebd74dcff639737