Introduce subnet pool prefix operations extension
This introduces a new extension that provides API's that enable
operations against the prefix list of a subnet pool to be more
explicit. In its current form, the subnetpool extension does not
clearly define how operations on the prefix list of a subnet pool
are to be performed and leaves unnecessary ambiguity to be handled
by implementers of the extension. This extension makes prefix
management for subnet pools explicit operations, while making support
for this discoverable for clients.

Change-Id: Ia84d767f582f28a2fc77799a58f6729a7fd42f34
Related-Bug: #1792901