Remove ml2's accidental dependency on l3
The accidental dependency was never in effect since the neutron side of
the relevant changes was not merged yet.

I think I made a mistake in https://review.openstack.org/631515.
We added the 'router' extension as a dependency of the
'floatingip-autodelete-internal' extension. Which looks a perfectly
reasonable thing to do at first sight. However since the 'external-net'
extension was de-extensionalized and made part of the ml2 plugin, the
'floatingip-autodelete-internal' extension also had to be implemented by
the ml2 plugin. This complicated setup practically made the l3 plugin
a dependency of the ml2 plugin. (That's why unit tests started failing
in patch set #3 of the neutron change.) Which of course is non-sense.

So this change removes the dependency. The neutron side of this
change still degrades gracefully even without the explicit dependency
between the extensions, so I don't think we're losing anything by not
having that dependency.

Change-Id: I8825eaf4f46ea2639131e34f9b833af1de6ab1b4
Needed-By: https://review.openstack.org/624751
Partial-Bug: #1806032
Related-Change: https://review.openstack.org/631515