fix WarningsFixture for public consumption
The neutron.tests.tools was recently made public and along with it the
WarningsFixture. However the WarningsFixture as-is doesn't support
filtering warnings from other modules. This patch address that by
accepting an constructor arg to add other module regular expressions to
filter. It also moves WarningsFixture to the fixture module where it
seems to make the most sense.

For a sample consumption patch see:
https://review.openstack.org/#/c/651370/

Change-Id: I08c713916aef407ba108ed2c67c3e99892f5883a