remove use of legacy modules
The notion of legacy modules has been in neutron-lib from the start and
the intention was to place them under neutron_lib.legacy.*
However we've been working on neutron-lib for a few years now and have
never used legacy modules. IMO it's safe to assume we won't use them
in neutron-lib.

This patch removes the empty sub-package neutron_lib.legacy and removes
the discussion of them from the docs.

Change-Id: Ie7a4cf65e37c3f2dcd3446232dba9695f41e314d