update db fixtures for consumption testing
Today neutron-lib has a SqlFixture that's public, but not used by
anyone else. However, this fixture doesn't contain the same
functionality as the fixture in neutron; so neutron breaks when
using it as-is today.

This patch preps for consumption of the SqlFixture(s) by:
- Moving a few more of SqlFixtures from neutron into lib; these are used
by consumers today.
- Updates the existing SqlFixture to include some functionality that
neutron requires.
- Makes all these SqlFixtures private. This is temporary so that we can
have a little time to test them in neutron and other's gates while
making sure no one uses them from lib. Once we are done testing we
will update them again in lib to be public and consume them. At that
point some UTs can also be added with a release note.

Neutron sample consumption patch:
https://review.openstack.org/#/c/651907/

Change-Id: Iaefc83d852a8dca48569e543f535f7c49f5ad0db