Update Python 3 test runtimes for Train
This goal is to implement the process set out in the 2018-10-24 Python
Update Process TC resolution[1], for the Train cycle to ensure unit
testing is in place for all of the Tested Runtimes for Train[2].
In practice, this generally means adding unit tests for Python 3.7 and
dropping unit tests for Python 3.5. Using the Zuul template for Train
will ensure that all projects that support Python3 will be tested
against the agreed runtime versions, and make it easier to update them
in future.

Specifically this patch:
- Uses the openstack-python3-train-jobs zuul template rather than
templates for python 36/37.
- Adds Python 3.7 to the setup.cfg classifier.
- Adds py37 to the envlist of tox.ini.

[1]https://governance.openstack.org/tc/resolutions/20181024-python-update-process.html
[2]https://governance.openstack.org/tc/reference/runtimes/train.html

Change-Id: I92e07412e06d5e0bcd4a3ef63643d140243f8205
Depends-On: https://review.opendev.org/#/c/641878/