sync sql fixtures from neutron
This patch updates the sql fixtures in neutron_lib.fixtures to
match their implementation in neutron. This is necessary due to some
issues with the existing fixtures in neutron-lib. The fixtures are
still marked private as we'll need to test them again before
making them public for consumption.

Sample consumption patch: https://review.opendev.org/#/c/657887/

Change-Id: I3e1ed2374125f2ac1e57db70bf206fdbeb3f82e4