Only have one number to name mapping for IPv6 ICMP
Neutron normalized the IPv6 ICMP protocol string to 'ipv6-icmp',
but the number to name mapping dict incorrectly shows 'icmpv6':

>>> constants.IP_PROTOCOL_NUM_TO_NAME_MAP['58']
'icmpv6'

This is because in the name to number mapping dict there are
two keys, 'ipv6-icmp' and 'icmpv6', that have a value of 58
in order to keep backwards-compatibility. Since the number
to name mapping is built automatically from that, it overwrites
the '58' key value to be 'icmpv6'. Filter it out so there is
only a single mapping to 'ipv6-icmp'.

Change-Id: I6b6c34c13f006ffca4cb4e2f18a2f825764374ec
Related-bug: #1582500