Add ethertype validator for custom ethertype validation
Since the restriction on ethertypes is being raised so that ethertypes
other than IPv4 and IPv6 can be specified in security group rules, add a
validator to ensure that ethertypes will be a valid two octet value.
The ethertype handling is configurable so make sure the validator
handles both possible configuration alternatives: the original style
where only IPv4 and IPv6 are valid, and the new style where any valid
two octet number (or it's string equivalent) works.

Related-Bug: #1832758
Needed-By: https://review.opendev.org/670203
Change-Id: I900f370f1de20b5137f138e54c5eec2102f77cce