Add retry and debug logs to NoAuthClient of placement client
Based on logstash (see [0])
test_configurations_are_synced_towards_placement sporadically fails.
Based on the logs it seems that when the dummy placement wsgi server
restarts the client can't connect to it, to make the situation cleaner
I added some logs for the different possible requests exceptions and a
simple retry mechanism.

[0]: http://logstash.openstack.org/#dashboard/file/logstash.json?query=message%3A%5C%22test_agent_bandwidth_report.py*%20line%20249%2C%20in%20test_configurations_are_synced_towards_placement%5C%22

Change-Id: I283c4b159ad0b9ffef8a6f5a3bfe1fe2feac8210