Change the requirements for the subnet_dns_publish_fixed_ip extension
During the implementation of the subnet_dns_publish_fixed_ip
extension [0] it was noticed that it will be better to implement just
a single stack of dns_integration extensions instead of having the
option of combining them flexibly. So we need to have the new extension
require the top of the current stack instead of just the basic extension.

[0] https://review.opendev.org/#/c/662409/

Change-Id: Ie0cccc5bd736b0302f956a7f01a9c16a8cc75200