Extend qos DriverBase with validate_rule_for_port()
The openvswitch qos driver supports the min-bw qos rule only on
networks backed by physnets. Currently the qos driver interface does
not have a method that takes both the policy/rule and the network/port,
therefore we don't have enough information in the driver to tell if
we support the min-bw rule for a port or not. This change introduces
validate_rule_for_port(..., rule, port) so we have all the information
to tell.

The base method validate_rule_for_port() is concrete and returns True
for backwards compatibility.

Change-Id: I58a47bb895bede219cfeebbf0665aaf01660ce89
Needed-By: https://review.opendev.org/705695
Partial-Bug: #1861442
Related-Bug: #1819029