[OVN] Add distributed device_owner
Add distributed device owner for services like OVN
Metadata and OVN DHCP.

OVN Metadata uses 'localport' port concept in OVN [1]:

"A connection to a local VIF. Traffic that  arrives  on  a
localport  is  never  forwarded  over a tunnel to another
chassis. These ports are present  on  every  chassis  and
have  the  same  address  in all of them. This is used to
model connectivity to local services that  run  on  every
hypervisor." [1]

For OVN DHCP we use localport fixed-ip allocation to get
IP address used as `server_id`, but the DHCP port is not bound
anywhere.

Because of those two reasons we should define new device_owner
and treat it differently in Neutron API, to unlock the OVN
distributed services benefits and to not use the same port_type
for Neutron non-distributed services.

[1] https://www.ovn.org/support/dist-docs/ovn-nb.5.txt

Change-Id: I53828ec6e1f0d2861eebfadcb9d46551fbf1a6af