Change policy on neutron-lib migration
Per the results of discussion at the Virtual PTG in June 2020, we are
changing our policy with regards to neutron-lib code forklifts.  We are
signaling that the mass migration of code from neutron to neutron-lib
will no longer be actively pursued, and simplifying the process for such
migrations - should they become necessary - by using debtcollector
removals.

Change-Id: I5669caf7e579c066acc52887a7f31ee8e238093f