[api-ref] Added warning about overwriting router's default routes
When user will add to the router extra route which will overwrite
default router of one of the subnets attached to the router, and later
will remove this extra route, traffic to/from such subnet will be
broken.
This patch adds warning about that to the router's extra route API
documentation.

Co-Authored-By: Brian Haley <haleyb.dev@gmail.com>

Change-Id: I488530f6805392be7de205d7a55081b053991c78