Add desc support to the BaseWorker initialization
In some cases, when we call the "BaseWorker.start" method (eg: use
oslo_service.service.ProcessLanuch to pull up the worker), we cannot
pass the "desc" parameter, so we need to pass in the desc parameter
during initialization.

this patch move the "desc" parameter to the initialization, when call
"start" method, if set "desc", it will be used first. if not, "desc"
passed in initialization are used. this design solves the problem and
keeps compatibility.

Partial-bug: #1910623

Change-Id: I4df8c319308224175a7f735cfc966b1080caa227