Populate full VLAN range in "parse_network_vlan_ranges"
The method "parse_network_vlan_ranges" parses the
"network_vlan_ranges" parameter that defines, for the VLAN type
driver, the available physical networks and the assigned VLAN
ID ranges.

If only the network name is passed without defining a VLAN ID range,
that means all valid VLAN IDs are allowed (1, 4094).

The returned variable should contain this range in this case.

Change-Id: I53726d64167ba38c233c2cfc095ff6462bcb664e
Closes-Bug: #1918274