Remove warning about session with and without enginefacade
Now we moved all code to the new enginefacade so we don't need this
check anymore.
And BTW. I think it wasn't working properly as even if we are using new
facade we still see plenty of those warnings in e.g. logs of the
neutron functional tests CI job.

Partially-Implements blueprint: enginefacade-switch

Change-Id: Ib2933e4f52e8d7585c13a0d1a3485389fa37b432