Return one single model field in "get_collection_query"
Added a new input parameter, "field". This parameter represents a
column from the model DB register. The "query_with_hooks" method
will return a SQL query that retrieves only the selected field
instead of the full DB model, that implies all register back
references (that implies JOIN clauses for each child DB register
related to the parent DB model).

The goal of this patch is to reduce the query complexity and thus
reduce the DB execution time.

This new parameter is a string.

Related-Bug: #1925528

Change-Id: Ia2f02ed576fda9d91367619de35d7c0900336a4e