Allow PUT on direction param in QoS minimum pps rule
Based on discussion in the previous review [1] and IRC [2], it was
decided to allow direction parameter in QoS minimum packet rate
rule to be updated with a PUT request. This behavior matches current
behavior of other QoS rules.

If the rule to be updated is associated with a bound port, the request
is going to be rejected by the server.

Since direction-oriented and direction-less QoS min pps rules are
mutually exclusive, the request is going to be rejected if it would
result in mixing direction-oriented and direction-less rules in a single
policy.

[1] https://review.opendev.org/802486
[2] https://meetings.opendev.org/irclogs/%23openstack-neutron/latest.log.html#t2021-08-09T07:56:39

Related-Bug: #1922237
See-Also: https://review.opendev.org/785236
Change-Id: If1f7b28184c740c30388f564b704ba42c8967527