Rephrase QoSRuleParameterConflict message
The name of QoSRuleParameterConflict exception class suggests that
it's a generic exception that can be used by multiple QoS rules to
raise a parameter conflict. On the other hand, exception's message
refers to bandwidth which could be confusing when used with QoS
minimum packet rate rule.

Related-Bug: #1922237
See-Also: https://review.opendev.org/785236
Change-Id: I05d33fd3337b755b1b0b8a5986739c76a1db3d84