Improve "get_collection_count" method
This improvement is based on [1]. "get_collection_count" can now
request a model query requesting only one single field from the
model. That simplifies the SQL query and improves the performance.
The result provided will have the same number of elements; that
means the result of the collection count will be the same.

[1]https://review.opendev.org/c/openstack/neutron-lib/+/787672

Closes-Bug: #1942863

Change-Id: Ie0bffac20b9c40933b64f163de3cc03d45b24682