Make update_qos_minbw_allocation() more generic
update_qos_minbw_allocation() function is used to update Placement
allocation. Initially, only minimum bandwidth rule allocated resources
in Placement, but with the introduction of minimum packet rate rule
that's changed. Because of that, we should rename the function to make
it more generic and to avoid confusion.

The second problem with this function is that it allows to update
resources only of a single RP at a time, even if multiple RPs are
associated with the same consumer UUID. With introduction of a
minimum packet rate rule it makes sense to allow to update resources of
multiple RPs in a single API call. To accommodate this, we need to
slightly modify arguments this function takes, and embed RP UUID
in alloc_diff, rather than pass it as a separate parameter.

Addresses TODO comment from [1].

[1] https://opendev.org/openstack/neutron/src/branch/master/neutron/services/qos/qos_plugin.py#L68

Partial-Bug: #1943724
Related-Bug: #1922237
See-Also: https://review.opendev.org/785236
Change-Id: Ie28b95e8ed351ab88db1fc75c83a02c474582e0b