Add custom range parameters on port_range validator
To maintain the behavior of port forwarding
validations that denies the port number 0 when
creating a new port forwarding rule for floating
IPs, we will need to extend the
'validate_port_range_or_none' validator to allow
us to define the acceptable port ranges when
validating a value.

This patch is one of a series of patches
to implement floating ip port forwarding with
port ranges.

The specification is defined in:
https://github.com/openstack/neutron-specs/blob/master/specs/wallaby/port-forwarding-port-ranges.rst

Change-Id: I0e11885e15616509942d60fbced3ff8e14cac654
Implements: blueprint floatingips-portforwarding-ranges
Related-Bug: #1885921