Rehome ovsfw constants and utils modules
Constants module contains definition of the OF registry numbers used
by ovs firewall driver.
After [1] some other registries are also used to store information
related to QoS bw limit and min bw rules.
So it would be good to have list of all used registers in one place and
reuse them where it's needed. That's why this patch moves those
constants to the neutron_lib.

Additionally it also adds utils module with some helper functions which
are using those constants and are actually used not only in neutron but
also in neutron-fwaas.

[1] https://review.opendev.org/c/openstack/neutron/+/832662

Related-Bug: #1959567
Change-Id: I89cc83699770e97bba369f50dd4d9e0ed6563aa4