Check proper config option to see if scope is enforced or not
Oslo policy introduced 2 config options which are related to the new
"Secure RBAC" policies and are in most cases used together:
* enforce_new_defaults
* enforce_scope

When I initially worked on new default policies in Neutron I always
set both of them to True together and I wrongly added checking
if enforce_new_defaults is set to True to check or set system_scope
parameter in context object.
As enforce_scope is related to scope actually, it should be set/check
when enforce_scope is enabled, not when enforce_new_defaults is enabled.

This patch fixes that by checking proper config option.

Change-Id: Ib5d4134bf59ea37d983f33807b396b31ffd9719d