Add floating-ip-port-forwarding-detail shim extension
if we list floating ip and want to operate a port forwarding, we cannot
call the update 'port forwarding' api, because we don't know the port
forwarding id.
The reason for exposing 'internal_port_id' is also to find the
associated port as soon as possible, so as to facilitate users to obtain
vm information and decide whether to update this port forwarding.
this patch adds the port forwarding returned contents: 'id' and
'internal_port_id' when list floatingip.

Partial-bug: #1971646

Change-Id: I34aa6948fccc669fc729d032b540bd1aa0c17ce9