[sqlalchemy-20] Use session.transaction information to decide if active
The "session.transaction" member "_connections" is a dictionary with the
active database connections of the current transaction. Neutron will use
it to determine if the session is active or not. This parameter will
provide more accurate information than "new", "dirty" or "deleted". A
read (SELECT) operation won't be included on those lists while the
transaction is currently active.

This patch is patch is a direct transcription of [1].

[1]https://review.opendev.org/c/openstack/neutron/+/843256

Related-Bug: #1975542

Change-Id: Ibc04a1940c3833d3e10dbfc011598ac11c58ba13