Skip test_qos_dscp_create_and_update API test in stable/wallaby
Fix for this test for branches which don't supports DSCP mark 44 is
in neutron-tempest-plugin [1] but it's not in version used for
stable/wallaby branch and there's no way to backport it
(neutron-tempest-plugin repo is branchless).
As stable/wallaby is already in EM phase we can simply skip this one
failing test in that branch to unblock the gate.

[1] https://review.opendev.org/c/openstack/neutron-tempest-plugin/+/854369

Related: #1987378
Change-Id: I16ec53356ec3c5daa1dfa192a1cfa01816a02598