New ``network-ha`` API definition
This new network API field, that can be used during the network
creation, will trigger the creation of a ``ha_router_networks``
database register. This register binds the project with the
created network and defines it as the high availability network
of the project, that is unique per project.

The default value is "False".

Related-Bug: #2016198

Change-Id: Id6e434060a7559026f9083904a91213b39361336