Introduce "HasProjectPrimaryUniqueKey" class
This class is used for the database model definitions. It adds the
"project_id" field as primary key, non nullable and unique. This
is the only class that defined this field as unique; that implies
any register created will be unique per project.

For example, [1] is introducing this database limitation for the
"ha_router_networks" registers: only one HA network per project
can be created.

[1]https://review.opendev.org/c/openstack/neutron/+/881735/

Change-Id: Id26acbb29a8ddacf4e621e847f2cc05a7ae7f9d6
Related-Bug: #2016198