Context.elevated() method sets all required roles for context object
If context should be elevated, it should always have "admin", "member"
and "reader" roles set as admin user always have "member" and "reader"
role as well.
Usually, when context is created by keystone it is like that but in some
cases, e.g. when noauth middleware is used instead of keystone it's not
like that and then context from the environment don't have any role set
so we should make sure that elevated context have all required roles set
correctly.

Closes-Bug: #2019946
Change-Id: Ic70202d1b41ea64ffd63dc910b7852fe75421fa9