Add a "GROUP BY" clause on queries with RBAC entries
As reported in the Neutron patch [1], this change introduce a
"GROUP BY" clause on the SQL queries with RBAC entries. With [1],
all resouces with RBAC entries ('network', 'qospolicy',
'securitygroup', 'addressscope', 'subnetpool', 'addressgroup') will
load the RBAC entries with "joined" strategy.

Because of the low cardinality of the RBAC query when all the RBAC
registers are in one single project, this patch groups the resource
queries by the resource ID. That will reduce the results returned by
the SQL engine to only the singular registers required.

[1]https://review.opendev.org/c/openstack/neutron/+/884877

Related-Bug: #1918145
Change-Id: I800e0356714d59ba93ab6252c77be0a82f024055