Add short warning about UUIDs in the SG rule templates api
SG rule template API accepts any string (any uuid) as remote group id or
remote address group id. It may be something what's not existing in the
time of the creation of template and that's fine. But if it will be not
existing when regular SG will be created for project it will fail making
SG for that project.
This patch adds small warning about this to the default SG rules
template api-ref.

Additionally this patch updated some small issues in the "create
default security group rule" section.

Related-Bug: #1983053
Change-Id: I9bad07d8ed11796047883f87b45a7da0799cdcb1