Update default for BFD/ECMP router extra attributes
During review of the series it was requested to make the default
values for 'enable_default_route_bfd' and
'enable_default_route_ecmp' configurable.

In order to determine the correct value at runtime, Neutron needs
the attributes to be filled with something other than 'True' or
'False' when not provided in the request.

Needed-By: I581f4c5e5cfa275f8a6f0adec405f205e877ac55
Related-Bug: #2002687
Change-Id: I9096685fb79a84e11a8547a5aaa16f7f2df48a56